--- packages/scripts/src/tasks/providers/android.ts
@@ -1,15 +1,0 @@
-const add_assets = () => null;
-const set_launcher_icon = () => null;
-const set_splash_image = () => null;
-const set_app_meta = () => null;
-
-const generate_splash_image = () => {
-  return `cordova-res android --type splash  --skip-config --copy`;
-};
-
-export default {
-  // add_assets,
-  // set_launcher_icon,
-  // set_splash_image,
-  // set_app_meta,
-};
